Api key added to cms routes

+++ index.ts
@@ -96,7 +96,11 @@ app.get('/units', async (req, res) => {
 });
 
 app.post('/cms', async (req: any, res) => {
-  return await getPerformanceForCoreUnits(req.body.inputData, 1000, 1000);
+  if (req.headers['x-api-key'] == process.env.CMS_API_KEY) {
+    return await getPerformanceForCoreUnits(req.body.inputData, 1000, 1000);
+  } else {
+    res.code(401).send(new Error('You are not authorized'));
+  }
 });
 
 app.get('/preparedComps', (req, res) => {
@@ -109,15 +113,19 @@ app.get('/preparedComps', (req, res) => {
 
 app.get('/cms/comps', async (req, res) => {
   try {
-    const comps = await prisma.compositionJSON.findMany();
-    return comps.map((comp) => {
-      const result = {
-        id: Number(comp.id),
-        json: comp.json,
-        visibility: comp.visibility
-      };
-      return result;
-    });
+    if (req.headers['x-api-key'] == process.env.CMS_API_KEY) {
+      const comps = await prisma.compositionJSON.findMany();
+      return comps.map((comp) => {
+        const result = {
+          id: Number(comp.id),
+          json: comp.json,
+          visibility: comp.visibility
+        };
+        return result;
+      });
+    } else {
+      res.code(401).send(new Error('You are not authorized'));
+    }
   } catch (error: any) {
     return { error: error.message };
   }
@@ -125,11 +133,15 @@ app.get('/cms/comps', async (req, res) => {
 
 app.post('/cms/changeVisibility', async (req: any, res) => {
   try {
-    await prisma.compositionJSON.update({
-      where: { id: req.body.id },
-      data: { visibility: req.body.visibility }
-    });
-    return { info: 'visibility updated' };
+    if (req.headers['x-api-key'] == process.env.CMS_API_KEY) {
+      await prisma.compositionJSON.update({
+        where: { id: req.body.id },
+        data: { visibility: req.body.visibility }
+      });
+      return { info: 'visibility updated' };
+    } else {
+      res.code(401).send(new Error('You are not authorized'));
+    }
   } catch (error: any) {
     return { error: error.message };
   }
@@ -137,10 +149,14 @@ app.post('/cms/changeVisibility', async (req: any, res) => {
 
 app.delete('/cms/comps/:id', async (req: any, res) => {
   try {
-    await prisma.compositionJSON.delete({
-      where: { id: req.params.id }
-    });
-    return { info: 'composition deleted' };
+    if (req.headers['x-api-key'] == process.env.CMS_API_KEY) {
+      await prisma.compositionJSON.delete({
+        where: { id: req.params.id }
+      });
+      return { info: 'composition deleted' };
+    } else {
+      res.code(401).send(new Error('You are not authorized'));
+    }
   } catch (error: any) {
     return { error: error.message };
   }
@@ -148,9 +164,13 @@ app.delete('/cms/comps/:id', async (req: any, res) => {
 
 app.post('/cms/save', async (req: any, res) => {
   try {
-    console.log(req.body.composition);
-    await saveCompositionIntoDatabase(req.body.composition);
-    return { info: 'data succesfully saved' };
+    if (req.headers['x-api-key'] == process.env.CMS_API_KEY) {
+      console.log(req.body.composition);
+      await saveCompositionIntoDatabase(req.body.composition);
+      return { info: 'data succesfully saved' };
+    } else {
+      res.code(401).send(new Error('You are not authorized'));
+    }
   } catch (e: any) {
     return { error: e.message };
   }
@@ -470,8 +490,7 @@ app.get('/augments-ranking/:stage', async (req: any, res) => {
 });
 
 app.get('/test', async (req: any, res) => {
-  await collectDataAboutRankings(1000);
-  console.log('done');
+  res.code(401).send(new Error('You are not authorized'));
 });
 
 app.get('/unit/:id', async (req: any, res) => {
